+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/global-121/121-platform/compare/v1.104.0...master)
 
+### Removed
+
+- All audio-features are removed from the PA-App. To prevent a lot of "local changes" during development,
+  make sure to remove any previously generated `*.webm`/`*.mp3` audio-files at: `/interfaces/PA-App/src/assets/<language-code>`
+
 ---
 
 ## [1.104.0](https://github.com/global-121/121-platform/compare/v1.103.3...v1.104.0) - 2023-10-30